Adding a new problem.
Code in one file and the outputs and respective times in another.
I almost can't believe how fast mine are compared to some of the others in the thread. I checked against some of the other soultions done in python with the same timer stuff I used and it seems fine so that's good I guess. :)